Move `fieldtype` under `tools.compiler`
