+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/fieldtype/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
+ * Copyright 2018, TeamDev Ltd. All rights reserved.
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -22,6 +22,6 @@
  * Classes for working with different proto field types.
  */
 @ParametersAreNonnullByDefault
-package io.spine.gradle.compiler.message.fieldtype;
+package io.spine.tools.compiler.fieldtype;
 
 import javax.annotation.ParametersAreNonnullByDefault;
