+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/rejection/RejectionWriter.java
@@ -28,8 +28,8 @@ import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.TypeName;
 import com.squareup.javapoet.TypeSpec;
 import io.spine.base.ThrowableMessage;
-import io.spine.gradle.compiler.message.fieldtype.FieldType;
-import io.spine.gradle.compiler.message.fieldtype.FieldTypeFactory;
+import io.spine.tools.compiler.fieldtype.FieldType;
+import io.spine.tools.compiler.fieldtype.FieldTypeFactory;
 import io.spine.tools.proto.FieldName;
 import io.spine.tools.proto.RejectionDeclaration;
 import org.slf4j.Logger;
