+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/validation/ClassNames.java
@@ -28,7 +28,7 @@ import io.spine.tools.compiler.MessageTypeCache;
 import java.util.Collection;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static io.spine.gradle.compiler.message.fieldtype.FieldTypes.trimTypeName;
+import static io.spine.tools.compiler.fieldtype.FieldTypes.trimTypeName;
 import static io.spine.tools.java.PrimitiveType.getWrapperClass;
 import static io.spine.tools.proto.ScalarType.getJavaTypeName;
 import static io.spine.util.Exceptions.newIllegalArgumentException;
