+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/validation/MapFieldMethodConstructor.java
@@ -26,7 +26,7 @@ import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.TypeName;
 import io.spine.base.ConversionException;
-import io.spine.gradle.compiler.message.fieldtype.MapFieldType;
+import io.spine.tools.compiler.fieldtype.MapFieldType;
 import io.spine.tools.proto.FieldName;
 import io.spine.validate.ValidationException;
 import org.slf4j.Logger;
