+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/validation/RepeatedFieldMethodConstructor.java
@@ -28,8 +28,8 @@ import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterizedTypeName;
 import com.squareup.javapoet.TypeName;
 import io.spine.base.ConversionException;
-import io.spine.gradle.compiler.message.fieldtype.FieldType;
 import io.spine.tools.compiler.MessageTypeCache;
+import io.spine.tools.compiler.fieldtype.FieldType;
 import io.spine.tools.proto.FieldName;
 import io.spine.tools.proto.ScalarType;
 import io.spine.validate.ValidationException;
