+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/validation/SingularFieldMethodConstructor.java
@@ -26,8 +26,8 @@ import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
 import io.spine.base.ConversionException;
-import io.spine.gradle.compiler.message.fieldtype.FieldType;
 import io.spine.tools.compiler.MessageTypeCache;
+import io.spine.tools.compiler.fieldtype.FieldType;
 import io.spine.tools.proto.FieldName;
 import io.spine.validate.ValidationException;
 import org.slf4j.Logger;
