+++ gradle-plugins/model-compiler/src/test/java/io/spine/tools/gradle/compiler/ValidatingBuilderGenPluginShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.gradle.compiler.validate;
+package io.spine.tools.gradle.compiler;
 
 import io.spine.tools.gradle.given.GradleProject;
 import org.junit.Rule;
